+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/StrictPhysicalTable.java
@@ -6,7 +6,7 @@ import com.yahoo.bard.webservice.data.config.names.DataSourceName;
 import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService;
-import com.yahoo.bard.webservice.table.availability.ConcreteAvailability;
+import com.yahoo.bard.webservice.table.availability.StrictAvailability;
 
 import java.util.Map;
 import java.util.Set;
@@ -14,12 +14,12 @@ import java.util.Set;
 import javax.validation.constraints.NotNull;
 
 /**
- * An implementation of Physical table that is backed by a single fact table.
+ * An implementation of Physical table that is backed by a single fact table and has intersect availability.
  */
-public class ConcretePhysicalTable extends SingleDataSourcePhysicalTable {
+public class StrictPhysicalTable extends SingleDataSourcePhysicalTable {
 
     /**
-     * Create a concrete physical table.
+     * Create a strict physical table.
      *
      * @param name  Name of the physical table as TableName, also used as data source name
      * @param timeGrain  time grain of the table
@@ -27,7 +27,7 @@ public class ConcretePhysicalTable extends SingleDataSourcePhysicalTable {
      * @param logicalToPhysicalColumnNames  Mappings from logical to physical names
      * @param metadataService  Datasource metadata service containing availability data for the table
      */
-    public ConcretePhysicalTable(
+    public StrictPhysicalTable(
             @NotNull TableName name,
             @NotNull ZonedTimeGrain timeGrain,
             @NotNull Set<Column> columns,
@@ -39,12 +39,12 @@ public class ConcretePhysicalTable extends SingleDataSourcePhysicalTable {
                 timeGrain,
                 columns,
                 logicalToPhysicalColumnNames,
-                new ConcreteAvailability(DataSourceName.of(name.asName()), metadataService)
+                new StrictAvailability(DataSourceName.of(name.asName()), metadataService)
         );
     }
 
     /**
-     * Create a concrete physical table, the availability on this table is built externally.
+     * Create a strict physical table with the availability on this table built externally.
      *
      * @param name  Name of the physical table as TableName, also used as fact table name
      * @param timeGrain  time grain of the table
@@ -52,12 +52,12 @@ public class ConcretePhysicalTable extends SingleDataSourcePhysicalTable {
      * @param logicalToPhysicalColumnNames  Mappings from logical to physical names
      * @param availability  Availability that serves interval availability for columns
      */
-    public ConcretePhysicalTable(
+    public StrictPhysicalTable(
             @NotNull TableName name,
             @NotNull ZonedTimeGrain timeGrain,
             @NotNull Set<Column> columns,
             @NotNull Map<String, String> logicalToPhysicalColumnNames,
-            @NotNull ConcreteAvailability availability
+            @NotNull StrictAvailability availability
     ) {
         super(
                 name,
@@ -69,7 +69,7 @@ public class ConcretePhysicalTable extends SingleDataSourcePhysicalTable {
     }
 
     /**
-     * Create a concrete physical table.
+     * Create a strict physical table.
      * The fact table name will be defaulted to the name and the availability initialized to empty intervals.
      *
      * @param name  Name of the physical table as String, also used as fact table name
@@ -81,7 +81,7 @@ public class ConcretePhysicalTable extends SingleDataSourcePhysicalTable {
      * @deprecated Should use constructor with TableName instead of String as table name
      */
     @Deprecated
-    private ConcretePhysicalTable(
+    private StrictPhysicalTable(
             @NotNull String name,
             @NotNull ZonedTimeGrain timeGrain,
             @NotNull Set<Column> columns,