+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/ThetaSketchIntersectionReportingResources.groovy
@@ -37,7 +37,7 @@ import com.yahoo.bard.webservice.druid.util.FieldConverterSupplier
 import com.yahoo.bard.webservice.druid.util.ThetaSketchFieldConverter
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
 import com.yahoo.bard.webservice.table.Column
-import com.yahoo.bard.webservice.table.ConcretePhysicalTable
+import com.yahoo.bard.webservice.table.StrictPhysicalTable
 import com.yahoo.bard.webservice.table.LogicalTable
 import com.yahoo.bard.webservice.table.PhysicalTable
 import com.yahoo.bard.webservice.table.TableGroup
@@ -155,7 +155,7 @@ class ThetaSketchIntersectionReportingResources extends Specification {
 
         columns.add(lmc)
 
-        PhysicalTable physicalTable = new ConcretePhysicalTable(
+        PhysicalTable physicalTable = new StrictPhysicalTable(
                 "NETWORK",
                 DAY.buildZonedTimeGrain(UTC),
                 columns,